Extract shared NoiseConditionedModule base class
Move the noise conditioning, positional embedding, and label-position
interaction logic into a shared NoiseConditionedModule in
noise_conditioned.py. NoiseConditionedSFNO now subclasses it, overriding
_generate_noise to support isotropic noise via SHT. LocalNet uses the
base class directly with gaussian noise.
