+++ fme/ace/registry/test_stochastic_sfno.py
@@ -4,12 +4,9 @@ import pytest
 import torch
 from torch_harmonics import InverseRealSHT
 
-from fme.ace.registry.stochastic_sfno import (
-    Context,
-    NoiseConditionedSFNO,
-    isotropic_noise,
-)
+from fme.ace.registry.stochastic_sfno import NoiseConditionedSFNO, isotropic_noise
 from fme.core.device import get_device
+from fme.core.models.conditional_sfno.layers import Context
 
 
 @pytest.mark.parametrize("nlat, nlon", [(8, 16), (64, 128)])
